console_conf/subiquity: split identitymodel, as it is not the same at all.
