+++ subiquitycore/models/__init__.py
@@ -13,6 +13,5 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .identity import IdentityModel
 from .network import NetworkModel
-__all__ = ['IdentityModel', 'NetworkModel']
+__all__ = ['NetworkModel']
